refactor: 관리자 큐레이션 조회 위치 수정
