+++ src/main/java/net/pointofviews/curation/controller/specification/CurationAdminSpecification.java
@@ -11,6 +11,7 @@ import net.pointofviews.common.dto.BaseResponse;
 import net.pointofviews.curation.domain.CurationCategory;
 import net.pointofviews.curation.dto.request.CreateCurationRequest;
 import net.pointofviews.curation.dto.response.ReadCurationListResponse;
+import net.pointofviews.curation.dto.response.ReadCurationMoviesResponse;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -163,4 +164,39 @@ public interface CurationAdminSpecification {
     ResponseEntity<?> deleteCuration(
             @PathVariable Long curationId
     );
+
+
+
+    @Operation(summary = "관리자 모든 큐레이션 조회", description = "관리자가 생성한 모든 큐레이션을 조회합니다.")
+    @ApiResponses({
+            @ApiResponse(
+                    responseCode = "200",
+                    description = "조회 성공"
+            )
+    })
+    ResponseEntity<BaseResponse<ReadCurationListResponse>> adminReadAllCurations();
+
+    @Operation(summary = "관리자 큐레이션 상세 조회", description = "관리자가 큐레이션 상세 정보를 조회합니다.")
+    @ApiResponses({
+            @ApiResponse(
+                    responseCode = "200",
+                    description = "조회 성공"
+            ),
+            @ApiResponse(
+                    responseCode = "404",
+                    description = "조회 실패",
+                    content = @Content(
+                            mediaType = MediaType.APPLICATION_JSON_VALUE,
+                            examples = @ExampleObject(value = """
+                                    {
+                                      "message": "존재하지 않는 큐레이션 입니다."
+                                    }
+                                    """)
+                    )
+            )
+    })
+    ResponseEntity<BaseResponse<ReadCurationMoviesResponse>> adminReadCuration(
+            @PathVariable Long curationId
+    );
+
 }
