+++ src/main/java/net/pointofviews/curation/service/CurationAdminService.java
@@ -3,6 +3,7 @@ package net.pointofviews.curation.service;
 import net.pointofviews.curation.domain.CurationCategory;
 import net.pointofviews.curation.dto.request.CreateCurationRequest;
 import net.pointofviews.curation.dto.response.ReadCurationListResponse;
+import net.pointofviews.curation.dto.response.ReadCurationMoviesResponse;
 import net.pointofviews.curation.dto.response.ReadCurationResponse;
 
 public interface CurationAdminService {
@@ -10,4 +11,7 @@ public interface CurationAdminService {
     ReadCurationListResponse searchCurations(String theme, CurationCategory category);
     void updateCuration(Long curationId, CreateCurationRequest request);
     void deleteCuration(Long curationId);
+
+    ReadCurationListResponse readAllCurations();
+    ReadCurationMoviesResponse readCuration(Long curationId);
 }
