+++ src/main/java/net/pointofviews/curation/service/impl/CurationAdminServiceImpl.java
@@ -5,6 +5,7 @@ import net.pointofviews.curation.domain.Curation;
 import net.pointofviews.curation.domain.CurationCategory;
 import net.pointofviews.curation.dto.request.CreateCurationRequest;
 import net.pointofviews.curation.dto.response.ReadCurationListResponse;
+import net.pointofviews.curation.dto.response.ReadCurationMoviesResponse;
 import net.pointofviews.curation.dto.response.ReadCurationResponse;
 import net.pointofviews.curation.exception.CurationException;
 import net.pointofviews.curation.repository.CurationRepository;
@@ -14,6 +15,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
@@ -92,4 +94,40 @@ public class CurationAdminServiceImpl implements CurationAdminService {
         curationMovieRedisService.deleteAllMoviesForCuration(curationId);
     }
 
+
+    @Override
+    public ReadCurationListResponse readAllCurations() {
+        List<ReadCurationResponse> curationResponses = curationRepository.findAll()
+                .stream()
+                .map(curation -> new ReadCurationResponse(
+                        curation.getId(),
+                        curation.getTheme(),
+                        curation.getCategory(),
+                        curation.getTitle(),
+                        curation.getDescription(),
+                        curation.getStartTime()
+                ))
+                .collect(Collectors.toList());
+
+        return new ReadCurationListResponse(curationResponses);
+    }
+
+    @Override
+    public ReadCurationMoviesResponse readCuration(Long curationId) {
+        Curation curation = curationRepository.findById(curationId)
+                .orElseThrow(CurationException::CurationNotFound);
+
+        Set<Long> movieIds = curationMovieRedisService.readMoviesForCuration(curationId);
+
+        ReadCurationResponse curationResponse = new ReadCurationResponse(
+                curation.getId(),
+                curation.getTheme(),
+                curation.getCategory(),
+                curation.getTitle(),
+                curation.getDescription(),
+                curation.getStartTime()
+        );
+
+        return new ReadCurationMoviesResponse(curationResponse, movieIds);
+    }
 }
